use uint8 for dataset

+++ mnist.py
@@ -4,8 +4,7 @@ import pandas as pd
 IMAGE_SIZE = 28
 
 def load_train_data(data_path, validation_size=1000):
-    # train_data = np.genfromtxt(data_path, delimiter=',', dtype=np.float32)
-    train_data = pd.read_csv(data_path, header=None).values.astype(np.float32)
+    train_data = pd.read_csv(data_path, header=None, dtype=np.uint8).values.astype(np.float32)
     x_train = train_data[:, 1:] / 255
     y_train = train_data[:, 0]
     y_train = (np.arange(10) == y_train[:, None]).astype(np.float32)
@@ -27,7 +26,7 @@ def load_train_data(data_path, validation_size=1000):
     return x_train, x_val, y_train, y_val
 
 def load_test_data(data_path):
-    test_data = pd.read_csv(data_path, header=None).values.astype(np.float32)
+    test_data = pd.read_csv(data_path, header=None, dtype=np.uint8).values.astype(np.float32)
     x_test = test_data[:, 1:] / 255
 
     y_test = np.array(test_data[:, 0])
